+++ src/main/java/run/app/mapper/BlogMapper.java
@@ -11,7 +11,7 @@ public interface BlogMapper {
 
     int deleteByExample(BlogExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Blog record);
 
@@ -21,7 +21,8 @@ public interface BlogMapper {
 
     List<Blog> selectByUserExample(@Param("query") PostQueryParams postQueryParams, @Param("blogger_id") Integer blogger_id);
 
-    Blog selectByPrimaryKey(Integer id);
+
+    Blog selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") Blog record, @Param("example") BlogExample example);
